clean: binding.cc - moved each class in separate file
