+++ src/binding.cc
@@ -101,230 +101,6 @@ void init (Handle<Object> target) {
   target->Set(String::NewSymbol("assemble_document"), FunctionTemplate::New(AssembleDocument)->GetFunction());
 }
 
-Persistent<FunctionTemplate> TermGenerator::constructor_template;
-
-void TermGenerator::Init(Handle<Object> target) {
-  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
-  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-  constructor_template->SetClassName(String::NewSymbol("TermGenerator"));
-
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "set_database", SetDatabase);
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "set_flags", SetFlags);
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "set_stemmer", SetStemmer);
-
-  target->Set(String::NewSymbol("TermGenerator"), constructor_template->GetFunction());
-
-  Handle<Object> aO = constructor_template->GetFunction();
-  aO->Set(String::NewSymbol("FLAG_SPELLING"), Integer::New(Xapian::TermGenerator::FLAG_SPELLING), ReadOnly);
-}
-
-Handle<Value> TermGenerator::New(const Arguments& args) {
-  HandleScope scope;
-  if (args.Length())
-    return ThrowException(Exception::TypeError(String::New("arguments are ()")));
-  TermGenerator* that = new TermGenerator;
-  that->Wrap(args.This());
-  return args.This();
-}
-
-Handle<Value> TermGenerator::SetDatabase(const Arguments& args) {
-  HandleScope scope;
-  WritableDatabase* aDb;
-  if (args.Length() < 1 || !(aDb = GetInstance<WritableDatabase>(args[0])))
-    return ThrowException(Exception::TypeError(String::New("arguments are (Database)")));
-
-  TermGenerator* that = ObjectWrap::Unwrap<TermGenerator>(args.This());
-  try {
-    that->mTg.set_database(aDb->getWdb());
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-
-  return Undefined();
-}
-
-Handle<Value> TermGenerator::SetFlags(const Arguments& args) {
-  HandleScope scope;
-  if (args.Length() < 1 || !args[0]->IsInt32())
-    return ThrowException(Exception::TypeError(String::New("arguments are (integer)")));
-
-  TermGenerator* that = ObjectWrap::Unwrap<TermGenerator>(args.This());
-  try {
-    that->mTg.set_flags((Xapian::TermGenerator::flags)args[0]->Int32Value());
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-
-  return Undefined();
-}
-
-Handle<Value> TermGenerator::SetStemmer(const Arguments& args) {
-  HandleScope scope;
-  Stem* aSt;
-  if (args.Length() < 1 || !(aSt = GetInstance<Stem>(args[0])))
-    return ThrowException(Exception::TypeError(String::New("arguments are (Stem)")));
-
-  TermGenerator* that = ObjectWrap::Unwrap<TermGenerator>(args.This());
-  try {
-    that->mTg.set_stemmer(aSt->mStem);
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-
-  return Undefined();
-}
-
-Persistent<FunctionTemplate> Stem::constructor_template;
-
-void Stem::Init(Handle<Object> target) {
-  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
-  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-  constructor_template->SetClassName(String::NewSymbol("Stem"));
-
-  //NODE_SET_PROTOTYPE_METHOD(constructor_template, "set_database", SetDatabase);
-
-  target->Set(String::NewSymbol("Stem"), constructor_template->GetFunction());
-}
-
-Handle<Value> Stem::New(const Arguments& args) {
-  HandleScope scope;
-  if (args.Length() < 1 || !args[0]->IsString())
-    return ThrowException(Exception::TypeError(String::New("arguments are (string)")));
-  Stem* that;
-  try {
-    that = new Stem(*String::Utf8Value(args[0]));
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-  that->Wrap(args.This());
-  return args.This();
-}
-
-
-
-Persistent<FunctionTemplate> Query::constructor_template;
-
-void Query::Init(Handle<Object> target) {
-  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
-  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-  constructor_template->SetClassName(String::NewSymbol("Query"));
-
-  //NODE_SET_PROTOTYPE_METHOD(constructor_template, "fn", Fn);
-
-  target->Set(String::NewSymbol("Query"), constructor_template->GetFunction());
-
-  Handle<Object> aO = constructor_template->GetFunction();
-  aO->Set(String::NewSymbol("OP_AND"         ), Integer::New(Xapian::Query::OP_AND         ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_OR"          ), Integer::New(Xapian::Query::OP_OR          ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_AND_NOT"     ), Integer::New(Xapian::Query::OP_AND_NOT     ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_XOR"         ), Integer::New(Xapian::Query::OP_XOR         ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_AND_MAYBE"   ), Integer::New(Xapian::Query::OP_AND_MAYBE   ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_FILTER"      ), Integer::New(Xapian::Query::OP_FILTER      ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_NEAR"        ), Integer::New(Xapian::Query::OP_NEAR        ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_PHRASE"      ), Integer::New(Xapian::Query::OP_PHRASE      ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_VALUE_RANGE" ), Integer::New(Xapian::Query::OP_VALUE_RANGE ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_SCALE_WEIGHT"), Integer::New(Xapian::Query::OP_SCALE_WEIGHT), ReadOnly);
-  aO->Set(String::NewSymbol("OP_ELITE_SET"   ), Integer::New(Xapian::Query::OP_ELITE_SET   ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_VALUE_GE"    ), Integer::New(Xapian::Query::OP_VALUE_GE    ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_VALUE_LE"    ), Integer::New(Xapian::Query::OP_VALUE_LE    ), ReadOnly);
-  aO->Set(String::NewSymbol("OP_SYNONYM"     ), Integer::New(Xapian::Query::OP_SYNONYM     ), ReadOnly);
-}
-
-Handle<Value> Query::New(const Arguments& args) {
-  HandleScope scope;
-  int aN;
-  std::vector<std::string> aList;
-  for (aN = 1; aN < args.Length() && args[aN]->IsString(); ++aN)
-    aList.push_back(*String::Utf8Value(args[aN]));
-  if (args.Length() < 2 || !args[0]->IsInt32() || aN < args.Length())
-    return ThrowException(Exception::TypeError(String::New("arguments are (Query.op, string ...)")));
-  Query* that;
-  const char* aDesc;
-  try {
-  that = new Query((Xapian::Query::op)args[0]->Int32Value(), aList.begin(), aList.end());
-  aDesc = that->mQry.get_description().c_str();
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-  args.This()->Set(String::NewSymbol("description"), String::New(aDesc));
-  that->Wrap(args.This());
-  return args.This();
-}
-
-Persistent<FunctionTemplate> Document::constructor_template;
-
-void Document::Init(Handle<Object> target) {
-  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
-  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-  constructor_template->SetClassName(String::NewSymbol("Document"));
-
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "get_data", GetData);
-
-  target->Set(String::NewSymbol("Document"), constructor_template->GetFunction());
-}
-
-Handle<Value> Document::New(const Arguments& args) {
-  HandleScope scope;
-
-  if (args.Length() && !args[0]->IsExternal())
-    return ThrowException(Exception::TypeError(String::New("arguments are ()")));
-
-  Document* that = new Document(args.Length() ? (Xapian::Document*) External::Unwrap(args[0]) : new Xapian::Document);
-  that->Wrap(args.This());
-
-  return args.This();
-}
-
-Handle<Value> Document::GetData(const Arguments& args) {
-  HandleScope scope;
-
-  if (args.Length() < 1 || !args[0]->IsFunction())
-    return ThrowException(Exception::TypeError(String::New("arguments are (function)")));
-  GetData_data* aData;
-  try {
-    aData = new GetData_data(args.This(), Local<Function>::Cast(args[0]));
-  } catch (Local<Value> ex) {
-    return ThrowException(ex);
-  }
-
-  eio_custom(GetData_pool, EIO_PRI_DEFAULT, GetData_done, aData);
-
-  return Undefined();
-}
-
-int Document::GetData_pool(eio_req* req) {
-  GetData_data* aData = (GetData_data*) req->data;
-
-  try {
-  aData->data = aData->object->mDoc->get_data();
-  } catch (const Xapian::Error& err) {
-    aData->error = new Xapian::Error(err);
-  }
-
-  aData->poolDone();
-  return 0;
-}
-
-int Document::GetData_done(eio_req* req) {
-  HandleScope scope;
-
-  GetData_data* aData = (GetData_data*) req->data;
-
-  Handle<Value> argv[2];
-  if (aData->error) {
-    argv[0] = Exception::Error(String::New(aData->error->get_msg().c_str()));
-  } else {
-    argv[0] = Null();
-    argv[1] = String::New(aData->data.c_str());
-  }
-
-  tryCallCatch(aData->callback, aData->object->handle_, aData->error ? 1 : 2, argv);
-
-  delete aData;
-
-  return 0;
-}
-
 Persistent<FunctionTemplate> Mime2Text::constructor_template;
 
 void Mime2Text::Init(Handle<Object> target) {
