+++ src/query.cc
@@ -1,0 +1,50 @@
+#include "node-xapian.h"
+
+Persistent<FunctionTemplate> Query::constructor_template;
+
+void Query::Init(Handle<Object> target) {
+  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
+  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
+  constructor_template->SetClassName(String::NewSymbol("Query"));
+
+  //NODE_SET_PROTOTYPE_METHOD(constructor_template, "fn", Fn);
+
+  target->Set(String::NewSymbol("Query"), constructor_template->GetFunction());
+
+  Handle<Object> aO = constructor_template->GetFunction();
+  aO->Set(String::NewSymbol("OP_AND"         ), Integer::New(Xapian::Query::OP_AND         ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_OR"          ), Integer::New(Xapian::Query::OP_OR          ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_AND_NOT"     ), Integer::New(Xapian::Query::OP_AND_NOT     ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_XOR"         ), Integer::New(Xapian::Query::OP_XOR         ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_AND_MAYBE"   ), Integer::New(Xapian::Query::OP_AND_MAYBE   ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_FILTER"      ), Integer::New(Xapian::Query::OP_FILTER      ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_NEAR"        ), Integer::New(Xapian::Query::OP_NEAR        ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_PHRASE"      ), Integer::New(Xapian::Query::OP_PHRASE      ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_VALUE_RANGE" ), Integer::New(Xapian::Query::OP_VALUE_RANGE ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_SCALE_WEIGHT"), Integer::New(Xapian::Query::OP_SCALE_WEIGHT), ReadOnly);
+  aO->Set(String::NewSymbol("OP_ELITE_SET"   ), Integer::New(Xapian::Query::OP_ELITE_SET   ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_VALUE_GE"    ), Integer::New(Xapian::Query::OP_VALUE_GE    ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_VALUE_LE"    ), Integer::New(Xapian::Query::OP_VALUE_LE    ), ReadOnly);
+  aO->Set(String::NewSymbol("OP_SYNONYM"     ), Integer::New(Xapian::Query::OP_SYNONYM     ), ReadOnly);
+}
+
+Handle<Value> Query::New(const Arguments& args) {
+  HandleScope scope;
+  int aN;
+  std::vector<std::string> aList;
+  for (aN = 1; aN < args.Length() && args[aN]->IsString(); ++aN)
+    aList.push_back(*String::Utf8Value(args[aN]));
+  if (args.Length() < 2 || !args[0]->IsInt32() || aN < args.Length())
+    return ThrowException(Exception::TypeError(String::New("arguments are (Query.op, string ...)")));
+  Query* that;
+  const char* aDesc;
+  try {
+  that = new Query((Xapian::Query::op)args[0]->Int32Value(), aList.begin(), aList.end());
+  aDesc = that->mQry.get_description().c_str();
+  } catch (const Xapian::Error& err) {
+    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
+  }
+  args.This()->Set(String::NewSymbol("description"), String::New(aDesc));
+  that->Wrap(args.This());
+  return args.This();
+}
