+++ wscript
@@ -24,7 +24,8 @@ def build(bld):
   obj = bld.new_task_gen("cxx", "shlib", "node_addon")
   obj.cxxflags = ["-g", "-D_FILE_OFFSET_BITS=64", "-D_LARGEFILE_SOURCE", "-Wall"]
   obj.target = "xapian-binding"
-  obj.source = [ "src/enquire.cc", "src/database.cc", "src/writabledatabase.cc", "src/binding.cc" ]
+  obj.source = [ "src/binding.cc", "src/database.cc", "src/document.cc", "src/enquire.cc", 
+    "src/query.cc", "src/stem.cc", "src/termgenerator.cc", "src/writabledatabase.cc" ]
   #obj.lib = "xapian"
   obj.uselib = "XAPIAN PROFILER"
